+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Moved Document types in package mongo_db_query
 - Export Reorganization
+- Fixed issue with mongodb+srv connection String
 
 ## 0.1.0-1.1.dev